refactor: move unchanged translations into i18n module
