+++ src/room-card.js
@@ -12,6 +12,8 @@ import { evaluateAdaptiveImageConditions as evaluateAdaptiveConditions, getCondi
 
 import { normalizeAlertSensorConfig, isAlertSensorActive } from "./lib/alerts.js";
 
+import { TRANSLATIONS, getTranslation } from "./i18n/translations.js";
+
 const VERSION = "1.4.0";
 const EDITOR_DOM_REVISION = "6";
 const LOG_FLAG = `customCards_RoomCard_Logged_${VERSION}`;
@@ -409,314 +411,6 @@ if (!customElements.get("oneline-room-card-textfield")) {
 }
 
 // TRANSLATIONS
-const TRANSLATIONS = {
-  en: {
-    empty: "Empty", low: "Low", critical: "Critical", window: "Window", general: "General",
-    sensors_manual: "Sensors (Manual)", buttons: "Buttons", button: "Button", add_button: "Add Button",
-    main_climate: "Main Climate Device (Optional)", climate_info: "Fills Temp/Humidity automatically if empty below.",
-    temp_label: "Temperature (overrides climate)", target_temp_label: "Target Temperature", humid_label: "Humidity (overrides climate)", temp_unit: "Temperature Unit", temp_unit_default: "Home Assistant default",
-    window_label: "Windows (List)", battery_label: "Batteries (List)", name: "Name", icon: "Icon", color: "Icon Color",
-    humid_warn_threshold: "Humidity Warning Threshold (%)", high_humidity: "High humidity", device_unavailable: "Device unavailable",
-    force_color: "Force Manual Color (Always visible)", img_url: "Image URL", image: "Image", path: "Path (Tap Action)", entity: "Entity", device: "Device (Optional)",
-    image_entity: "Light Entity (Grayscale when off)", image_entity_help: "When the selected light/switch is off, the header image fades to grayscale.",
-    show_image: "Show background image",
-    image_preset_title: "Built-in room image", image_preset_help: "Choose a RoomCard image or keep using your own URL/upload.", image_preset_custom: "Own image",
-    image_preset_living_room: "Living room", image_preset_kitchen: "Kitchen", image_preset_bedroom: "Bedroom", image_preset_bathroom: "Bathroom",
-    image_preset_dining_room: "Dining room", image_preset_home_office: "Home office", image_preset_childrens_room: "Children's room", image_preset_hallway: "Hallway",
-    image_preset_guest_room: "Guest room", image_preset_garage: "Garage", image_preset_garden_patio: "Garden / patio", image_preset_balcony: "Balcony",
-    image_preset_basement: "Basement", image_preset_laundry_room: "Laundry room", image_preset_attic: "Attic", image_preset_workshop: "Workshop",
-    presence_sensor: "Presence Sensor (Motion/Person)", presence_detected: "Present", presence_chip_color: "Presence Chip Color", presence_solid_background: "Solid Presence Background",
-    area_setup: "Area Setup", area_setup_desc: "Automatically populate controls and sensors from area entities", area_picker: "Home Assistant Area", area_generate: "Generate from Area", area_no_entities: "No entities found in this area",
-    alert_sensors: "Alert Sensors", alert_sensor_add: "Add Alert Sensor", alert_sensor_entity: "Sensor", alert_sensor_above: "Above", alert_sensor_below: "Below", alert_sensor_state: "State", alert_border_color: "Card Border Color", alert_chip_collapsed: "Show as badges (collapsed)", active_alerts: "Active Alerts",
-    template: "Type Filter", add_template: "with Filter", add_prefix: "Add",
-    quick_add_title: "Quick Add",
-    quick_add_desc: "Quickly add buttons from existing entities.",
-    quick_add_entity_type_label: "Entity type",
-    quick_add_entity_label: "Entity (filtered)",
-    quick_add_entity_type_help: "Filters the entity list on the right.",
-    quick_add_entity_help: "Select an entity of this type and click ‘Add’.",
-    quick_add_entity_placeholder: "Select entity ({type})",
-    quick_add_empty_hint: "No matching entities found.",
-    quick_add_settings_caption: "Button settings (for added entries)",
-    expand_all: "Expand all",
-    collapse_all: "Collapse all",
-    label_position: "Label & Status Position",
-    label_position_all: "Label & status position (all buttons)",
-    use_global: "Use global setting",
-    pos_right: "Right",
-    pos_bottom: "Bottom",
-    pos_left: "Left",
-    pos_top: "Top",
-    row_type: "Row Type", type_entity: "Entity", type_template: "Template",
-    tmpl_content: "Content (Template)", tmpl_icon: "Icon (Template)", tmpl_color: "Color (Template)", tmpl_state: "State (Template)", tmpl_preview: "Preview",
-    tmpl_light: "Light", tmpl_switch: "Switch / Socket", tmpl_select: "Select", tmpl_climate: "Climate", tmpl_cover: "Cover / Shutter", tmpl_media: "Media Player",
-    show_state: "Show State", show_label: "Show Label", show_icon: "Show Icon", show_last_changed: "Last Changed", show_sparkline: "Show Sparkline", sparkline_detail: "Open history details", sparkline_hours: "History (hours)", sparkline_refresh: "Sparkline refresh (seconds)", sparkline_refresh_adjusted: "Enter 60–3600 seconds. The value was adjusted to {value}.", sparkline_detail_title: "History details", sparkline_loading: "Loading history…", sparkline_empty: "No history data available.", sparkline_error: "History could not be loaded.", sparkline_current: "Current", sparkline_min: "Min", sparkline_max: "Max", sparkline_average: "Average", lc_just_now: "just now", state_first: "State First", text_layout: "Text Order", primary_text: "First line", primary_state: "State / value first", primary_name: "Name first",
-    height: "Height", width: "Width", align: "Align", visible: "Visible", left: "Left", center: "Center", right: "Right",
-    tap_action: "Tap Action", hold_action: "Hold Action", double_tap_action: "Double Tap Action",
-    actions: "Actions",
-    act_more: "Details (Default)", act_toggle: "Toggle", act_navigate: "Navigate", act_call_service: "Action (service)", act_none: "None",
-    service: "Service (domain.service)", service_data: "Service Data (JSON)", target_entity: "Target entity",
-    live_preview: "Live preview",
-    upload_btn: "Upload Image", uploading: "Uploading...", upload_success: "Image uploaded", upload_optimized: "Image optimized and uploaded",
-    upload_unsupported: "Choose a JPEG, PNG, or WebP image.", upload_too_large: "The source image exceeds the 20 MB limit.", upload_decode_error: "The image could not be decoded.", upload_failed: "Upload failed ({status}).",
-    image_focal_point: "Image focal point", image_focal_help: "Click or drag the marker to keep the important area visible.", image_center: "Center", image_horizontal: "Horizontal position", image_vertical: "Vertical position",
-    show_name: "Show Title", header_badges: "Badge", badge_add: "Add Info Entry", badge_label: "Label (optional)", badge_background: "Background (rgba)", standard_badge_background: "Standard Badge Background (rgba)", badge_auto_climate_btn: "Automatically add climate control button",
-    visibility: "Visibility", visibility_cond: "Conditional Visibility", vis_entity: "Condition Entity", vis_state: "Show if state is", vis_invert: "Invert Logic (Hide if state corresponds)",
-    migration_title: "Action Required",
-    migration_text: "Card renamed to <b>oneline-room-card</b> to avoid conflicts.<br>Please change <code>type: custom:room-card</code> to <code>type: custom:oneline-room-card</code> in your YAML.",
-    control_mode: "Control Mode", ctrl_default: "Default", ctrl_slider: "Inline Slider", ctrl_buttons: "Inline Buttons", ctrl_full: "Full Controls", ctrl_all_options: "All Options",
-    slider_mode: "Slider Mode", slider_mode_brightness: "Brightness", slider_mode_color_temp: "Color Temperature",
-    slider_style: "Slider Style", style_inline: "Inline", style_bg: "Background",
-    collapsible: "Collapsible", default_state: "Default State", state_expanded: "Expanded", state_collapsed: "Collapsed",
-    header_height: "Header Height (px)",
-    typography: "Header Typography", name_font: "Room Name Font", info_font: "Info Line Font", show_header_text_shadow: "Header Text Shadow",
-    font_size: "Size (px)", font_weight: "Weight", font_style: "Style", font_color: "Color", badge_bg: "Badge Background",
-    card_behavior: "Card Behavior", behavior: "Collapse Mode", behavior_fixed: "Disabled", behavior_collapsed: "Collapsed", behavior_expanded: "Expanded", behavior_remember: "Remember",
-    header: "Header", configuration: "Configuration",
-    color_map: "State Colors", color_map_add: "Add State Color", color_map_state: "State",
-    window_always_show: "Always Show (incl. closed)", window_open_color: "Open Color", window_closed_color: "Closed Color",
-    window_solid_background: "Solid Background", window_labels: "Window Labels",
-    window_custom_label: "Custom Label",
-    window_open_states: "Open States (comma-separated)", window_state_colors: "State Colors", window_state_colors_add: "Add State Color",
-    sensors: "Sensors", show_chip_shadow: "Sensor Chip Shadow", show_status_border: "Status border and glow", show_status_border_help: "Controls only the warning outline and glow; sensor badges remain visible.",
-    icon_size: "Icon Size", global_icon_size: "Global Icon Size (px)",
-    header_info_offset: "Info Line Position",
-    header_name_offset: "Title Position",
-    header_sync_offsets: "Synchronize Positions",
-    global_button_bg: "Global Button Background",
-    button_bg: "Button Background Color",
-    show_cover_presets: "Position Presets",
-    cover_presets_label: "Preset Values (comma-separated)",
-    show_climate_presets: "Temperature Presets",
-    climate_presets_label: "Temperatures (comma-separated)",
-    show_hvac_modes: "HVAC Mode Chips", show_fan_modes: "Fan Speed Chips",
-    show_brightness_presets: "Brightness Presets",
-    brightness_presets_label: "Brightness values (comma-separated)",
-    show_brightness_value: "Show Brightness %",
-    show_color_favorites: "Color Favorites",
-    color_favorites_label: "Colors ('#hex' or 'r,g,b', comma-separated)",
-    show_media_sources: "Source Chips",
-    show_media_sound_modes: "Sound Mode Chips",
-    show_media_title: "Media Title",
-    sub_chips: "Sub-Chips", chip_add: "Add Chip", chip_entity: "Entity", chip_attribute: "Attribute (optional)", chip_icon: "Icon (optional)", chip_label: "Label (optional)", chips_position: "Chip Position", chips_top: "Above title", chips_bottom: "Below title",
-    vis_add: "Add Condition", vis_eq: "State is equal", vis_neq: "State is not equal", vis_above: "State is strictly greater than", vis_below: "State is strictly less than",
-    info_line_position: "Info Line Position", info_position_header: "Inside header (default)", info_position_below: "Below header",
-    last_activity_title: "Last Activity", last_activity_show: "Show last activity",
-    room_modes: "Room Modes", room_modes_help: "Run scenes or scripts directly from the room card.", room_mode_add: "Add Room Mode", room_mode_entity: "Scene or script", room_mode_active_when: "Active when", room_mode_remove: "Remove mode", room_mode_up: "Move mode up", room_mode_down: "Move mode down",
-    adaptive_images: "Adaptive images", adaptive_images_help: "The first matching rule replaces the fallback header image.", adaptive_image_add: "Add image rule", adaptive_image_name: "Rule name (optional)", adaptive_image_conditions: "Conditions", adaptive_image_preset: "Built-in image", adaptive_image_custom: "Custom URL / upload", adaptive_image_duplicate: "Duplicate image rule", adaptive_image_remove: "Remove image rule", adaptive_image_up: "Move image rule up", adaptive_image_down: "Move image rule down", adaptive_image_position: "Rule focal point",
-    status_groups: "Room status", status_groups_help: "Summarize configured entities without changing alert priority.", status_group_add: "Add status group", status_group_preset_lights: "Lights on", status_group_preset_windows: "Open windows", status_group_preset_media: "Active media", status_group_preset_power: "Power", status_group_entities: "Entities", status_group_active_states: "Active states (comma-separated)", status_group_display: "Display", status_group_count: "Count", status_group_value: "Numeric sum", status_group_unit: "Output unit", status_group_precision: "Decimals", status_group_hide_zero: "Hide when zero", status_group_details: "Open contributing entities", status_group_conditions: "Show group when", status_group_duplicate: "Duplicate status group", status_group_remove: "Remove status group", status_group_up: "Move status group up", status_group_down: "Move status group down", status_group_incompatible_units: "Incompatible units",
-    a11y_close: "Close", a11y_previous_track: "Previous track", a11y_play_pause: "Play or pause", a11y_next_track: "Next track",
-    a11y_mute: "Mute or unmute", a11y_volume: "Volume", a11y_expand: "Expand room controls", a11y_collapse: "Collapse room controls",
-    a11y_activate: "Activate {name}", a11y_set_value: "Set {name}", a11y_select_option: "Select {name}"
-  },
-  de: {
-    empty: "Leer", low: "Niedrig", critical: "Kritisch", window: "Fenster", general: "Allgemein",
-    sensors_manual: "Sensoren (Manuell)", buttons: "Buttons", button: "Button", add_button: "Button hinzufügen",
-    main_climate: "Haupt-Klima-Gerät", climate_info: "Füllt Temp/Feuchtigkeit automatisch, wenn unten leer.",
-    temp_label: "Temperatur (überschreibt Klima)", target_temp_label: "Soll-Temperatur", humid_label: "Luftfeuchtigkeit (überschreibt Klima)", temp_unit: "Temperatureinheit", temp_unit_default: "Home-Assistant-Standard",
-    window_label: "Fenster (Liste)", battery_label: "Batterien (Liste)", name: "Name", icon: "Icon", color: "Iconfarbe",
-    humid_warn_threshold: "Feuchte-Warnschwelle (%)", high_humidity: "Hohe Luftfeuchtigkeit", device_unavailable: "Gerät nicht verfügbar",
-    force_color: "Manuelle Farbe erzwingen (Immer sichtbar)", img_url: "Bild URL", image: "Bild", path: "Pfad (Tap Action)", entity: "Entität", device: "Gerät (Optional)",
-    image_entity: "Licht-Entität (Graustufen wenn aus)", image_entity_help: "Wenn die gewählte Licht-/Schalter-Entität aus ist, wird das Header-Bild in Graustufen dargestellt.",
-    show_image: "Hintergrundbild anzeigen",
-    image_preset_title: "Integriertes Raumbild", image_preset_help: "Wähle ein RoomCard-Bild oder nutze weiterhin eine eigene URL bzw. einen Upload.", image_preset_custom: "Eigenes Bild",
-    image_preset_living_room: "Wohnzimmer", image_preset_kitchen: "Küche", image_preset_bedroom: "Schlafzimmer", image_preset_bathroom: "Badezimmer",
-    image_preset_dining_room: "Esszimmer", image_preset_home_office: "Arbeitszimmer", image_preset_childrens_room: "Kinderzimmer", image_preset_hallway: "Flur",
-    image_preset_guest_room: "Gästezimmer", image_preset_garage: "Garage", image_preset_garden_patio: "Garten / Terrasse", image_preset_balcony: "Balkon",
-    image_preset_basement: "Keller", image_preset_laundry_room: "Waschküche", image_preset_attic: "Dachboden", image_preset_workshop: "Werkstatt",
-    presence_sensor: "Anwesenheits-Sensor (Bewegung/Person)", presence_detected: "Anwesend", presence_chip_color: "Farbe des Anwesenheits-Chips", presence_solid_background: "Durchgefärbter Anwesenheits-Chip",
-    area_setup: "Bereich-Setup", area_setup_desc: "Steuerungen und Sensoren automatisch aus dem Bereich übernehmen", area_picker: "Home Assistant Bereich", area_generate: "Aus Bereich generieren", area_no_entities: "Keine Entitäten in diesem Bereich gefunden",
-    alert_sensors: "Alarm-Sensoren", alert_sensor_add: "Alarm-Sensor hinzufügen", alert_sensor_entity: "Sensor", alert_sensor_above: "Über", alert_sensor_below: "Unter", alert_sensor_state: "Zustand", alert_border_color: "Kartenrahmenfarbe", alert_chip_collapsed: "Als Sammel-Badge anzeigen", active_alerts: "Aktive Alarme",
-    template: "Typ-Filter", add_template: "mit Filter", add_prefix: "Add",
-    quick_add_title: "Schnellerfassung",
-    quick_add_desc: "Schnell Buttons aus bestehenden Entitäten hinzufügen.",
-    quick_add_entity_type_label: "Entitätstyp",
-    quick_add_entity_label: "Entität (gefiltert)",
-    quick_add_entity_type_help: "Filtert die Entitätenliste rechts.",
-    quick_add_entity_help: "Wähle eine Entität dieses Typs und klicke anschließend auf ‘Hinzufügen’.",
-    quick_add_entity_placeholder: "Entität auswählen ({type})",
-    quick_add_empty_hint: "Keine passenden Entitäten gefunden.",
-    quick_add_settings_caption: "Button-Einstellungen (für hinzugefügte Einträge)",
-    expand_all: "Alle aufklappen",
-    collapse_all: "Alle zuklappen",
-    label_position: "Position von Label & Status",
-    label_position_all: "Label & Status Position (alle Buttons)",
-    use_global: "Globale Einstellung verwenden",
-    pos_right: "Rechts",
-    pos_bottom: "Unten",
-    pos_left: "Links",
-    pos_top: "Oben",
-    row_type: "Zeilentyp", type_entity: "Entität", type_template: "Template",
-    tmpl_content: "Text (Template)", tmpl_icon: "Icon (Template)", tmpl_color: "Farbe (Template)", tmpl_state: "Status (Template)", tmpl_preview: "Vorschau",
-    tmpl_light: "Licht", tmpl_switch: "Schalter / Steckdose", tmpl_select: "Auswahl", tmpl_climate: "Klima", tmpl_cover: "Rollladen / Abdeckung", tmpl_media: "Media Player",
-    show_state: "Status anzeigen", show_label: "Bezeichnung anzeigen", show_icon: "Icon anzeigen", show_last_changed: "Letzte Änderung", show_sparkline: "Sparkline anzeigen", sparkline_detail: "Verlaufsdetails öffnen", sparkline_hours: "Verlauf (Stunden)", sparkline_refresh: "Sparkline-Aktualisierung (Sekunden)", sparkline_refresh_adjusted: "Bitte 60–3600 Sekunden eingeben. Der Wert wurde auf {value} angepasst.", sparkline_detail_title: "Verlaufsdetails", sparkline_loading: "Verlauf wird geladen…", sparkline_empty: "Keine Verlaufsdaten verfügbar.", sparkline_error: "Verlauf konnte nicht geladen werden.", sparkline_current: "Aktuell", sparkline_min: "Min", sparkline_max: "Max", sparkline_average: "Durchschnitt", lc_just_now: "gerade eben", state_first: "Wert zuerst", text_layout: "Text-Reihenfolge", primary_text: "Erste Zeile", primary_state: "Wert zuerst", primary_name: "Name zuerst",
-    height: "Höhe", width: "Breite", align: "Ausrichtung", visible: "Sichtbar", left: "Links", center: "Mitte", right: "Rechts",
-    tap_action: "Antippen", hold_action: "Gedrückt halten", double_tap_action: "Doppelklick",
-    actions: "Aktionen",
-    act_more: "Details (Standard)", act_toggle: "Umschalten", act_navigate: "Navigieren", act_call_service: "Aktion (service)", act_none: "Nichts",
-    service: "Service (domain.service)", service_data: "Service Daten (JSON)", target_entity: "Ziel-Entität",
-    live_preview: "Live-Vorschau",
-    upload_btn: "Bild hochladen", uploading: "Wird hochgeladen...", upload_success: "Bild hochgeladen", upload_optimized: "Bild optimiert und hochgeladen",
-    upload_unsupported: "Bitte ein JPEG-, PNG- oder WebP-Bild wählen.", upload_too_large: "Das Ausgangsbild überschreitet das Limit von 20 MB.", upload_decode_error: "Das Bild konnte nicht gelesen werden.", upload_failed: "Upload fehlgeschlagen ({status}).",
-    image_focal_point: "Bildfokus", image_focal_help: "Klicke oder ziehe die Markierung auf den wichtigen Bildbereich.", image_center: "Zentrieren", image_horizontal: "Horizontale Position", image_vertical: "Vertikale Position",
-    show_name: "Titel anzeigen", header_badges: "Badge", badge_add: "Info-Eintrag hinzufügen", badge_label: "Bezeichnung (optional)", badge_background: "Hintergrund (rgba)", standard_badge_background: "Standard Badge Hintergrund (rgba)", badge_auto_climate_btn: "Klima-Steuerungs-Button automatisch hinzufügen",
-    visibility: "Sichtbarkeit", visibility_cond: "Bedingte Sichtbarkeit", vis_entity: "Bedingungs-Entität", vis_state: "Anzeigen falls Status gleich", vis_invert: "Logik umkehren (Ausblenden falls entsprechend)",
-    migration_title: "Handlung erforderlich",
-    migration_text: "Karte wurde in <b>oneline-room-card</b> umbenannt.<br>Bitte ändere <code>type: custom:room-card</code> zu <code>type: custom:oneline-room-card</code> in deiner YAML-Konfiguration.",
-    control_mode: "Steuerungsmodus", ctrl_default: "Standard", ctrl_slider: "Inline-Slider", ctrl_buttons: "Inline-Buttons", ctrl_full: "Alle Steuerungen", ctrl_all_options: "Alle Optionen",
-    slider_mode: "Slider Modus", slider_mode_brightness: "Helligkeit", slider_mode_color_temp: "Farbtemperatur",
-    slider_style: "Slider Stil", style_inline: "Inline", style_bg: "Hintergrund",
-    collapsible: "Einklappbar", default_state: "Standardzustand", state_expanded: "Ausgeklappt", state_collapsed: "Eingeklappt",
-    header_height: "Kopfzeilenhöhe (px)",
-    typography: "Header Typografie", name_font: "Raumname Schrift", info_font: "Info-Zeile Schrift", show_header_text_shadow: "Header-Textschatten",
-    font_size: "Größe (px)", font_weight: "Gewicht", font_style: "Stil", font_color: "Farbe", badge_bg: "Badge Hintergrund",
-    card_behavior: "Kartenverhalten", behavior: "Einklapp-Modus",
-    behavior_fixed: "Deaktiviert",
-    behavior_collapsed: "Eingeklappt",
-    behavior_expanded: "Ausgeklappt",
-    behavior_remember: "Remember",
-    header: "Header", configuration: "Konfiguration",
-    color_map: "Zustandsfarben", color_map_add: "Farbe hinzufügen", color_map_state: "Zustand",
-    window_always_show: "Immer anzeigen (auch geschlossen)", window_open_color: "Farbe geöffnet", window_closed_color: "Farbe geschlossen",
-    window_solid_background: "Durchgefärbter Hintergrund", window_labels: "Fensterbezeichnungen",
-    window_custom_label: "Eigene Bezeichnung",
-    window_open_states: "Geöffnete Zustände (kommagetrennt)", window_state_colors: "Zustandsfarben", window_state_colors_add: "Farbe hinzufügen",
-    sensors: "Sensoren", show_chip_shadow: "Sensor-Chip-Schatten", show_status_border: "Statusrahmen und Leuchten", show_status_border_help: "Steuert nur Warnrahmen und Leuchten; Sensor-Badges bleiben sichtbar.",
-    icon_size: "Icon-Größe", global_icon_size: "Globale Icon-Größe (px)",
-    header_info_offset: "Info-Zeile Position",
-    header_name_offset: "Titel Position",
-    header_sync_offsets: "Synchron Bewegen",
-    global_button_bg: "Globaler Button Hintergrund",
-    button_bg: "Button Hintergrund Farbe",
-    show_cover_presets: "Positions-Voreinstellungen",
-    cover_presets_label: "Voreinstellungen (kommagetrennt)",
-    show_climate_presets: "Temperatur-Voreinstellungen",
-    climate_presets_label: "Temperaturen (kommagetrennt)",
-    show_hvac_modes: "HVAC-Modus-Chips", show_fan_modes: "Lüftergeschwindigkeit-Chips",
-    show_brightness_presets: "Helligkeits-Voreinstellungen",
-    brightness_presets_label: "Helligkeiten (kommagetrennt)",
-    show_brightness_value: "Helligkeit % anzeigen",
-    show_color_favorites: "Lieblings-Farben",
-    color_favorites_label: "Farben ('#hex' oder 'r,g,b', kommagetrennt)",
-    show_media_sources: "Quellen-Chips",
-    show_media_sound_modes: "Soundmodus-Chips",
-    show_media_title: "Medientitel",
-    sub_chips: "Sub-Chips", chip_add: "Chip hinzufügen", chip_entity: "Entität", chip_attribute: "Attribut (optional)", chip_icon: "Icon (optional)", chip_label: "Bezeichnung (optional)", chips_position: "Chip-Position", chips_top: "Über dem Titel", chips_bottom: "Unter dem Titel",
-    vis_add: "Bedingung hinzufügen", vis_eq: "Zustand ist gleich", vis_neq: "Zustand ist nicht gleich", vis_above: "Numerisch größer als", vis_below: "Numerisch kleiner als",
-    info_line_position: "Info-Zeile Position", info_position_header: "Im Header (Standard)", info_position_below: "Unter dem Header",
-    last_activity_title: "Letzte Aktivität", last_activity_show: "Letzte Aktivität anzeigen",
-    room_modes: "Raum-Modi", room_modes_help: "Szenen oder Skripte direkt über die Raumkarte starten.", room_mode_add: "Raum-Modus hinzufügen", room_mode_entity: "Szene oder Skript", room_mode_active_when: "Aktiv wenn", room_mode_remove: "Modus entfernen", room_mode_up: "Modus nach oben", room_mode_down: "Modus nach unten",
-    adaptive_images: "Adaptive Bilder", adaptive_images_help: "Die erste passende Regel ersetzt das Standardbild im Header.", adaptive_image_add: "Bildregel hinzufügen", adaptive_image_name: "Regelname (optional)", adaptive_image_conditions: "Bedingungen", adaptive_image_preset: "Integriertes Bild", adaptive_image_custom: "Eigene URL / Upload", adaptive_image_duplicate: "Bildregel duplizieren", adaptive_image_remove: "Bildregel entfernen", adaptive_image_up: "Bildregel nach oben", adaptive_image_down: "Bildregel nach unten", adaptive_image_position: "Bildfokus der Regel",
-    status_groups: "Raumstatus", status_groups_help: "Konfigurierte Entitäten zusammenfassen, ohne die Alarmpriorität zu verändern.", status_group_add: "Statusgruppe hinzufügen", status_group_preset_lights: "Lichter an", status_group_preset_windows: "Offene Fenster", status_group_preset_media: "Aktive Medien", status_group_preset_power: "Leistung", status_group_entities: "Entitäten", status_group_active_states: "Aktive Zustände (kommagetrennt)", status_group_display: "Anzeige", status_group_count: "Anzahl", status_group_value: "Numerische Summe", status_group_unit: "Ausgabeeinheit", status_group_precision: "Nachkommastellen", status_group_hide_zero: "Bei null ausblenden", status_group_details: "Beitragende Entitäten öffnen", status_group_conditions: "Gruppe anzeigen wenn", status_group_duplicate: "Statusgruppe duplizieren", status_group_remove: "Statusgruppe entfernen", status_group_up: "Statusgruppe nach oben", status_group_down: "Statusgruppe nach unten", status_group_incompatible_units: "Nicht kompatible Einheiten",
-    a11y_close: "Schließen", a11y_previous_track: "Vorheriger Titel", a11y_play_pause: "Wiedergabe oder Pause", a11y_next_track: "Nächster Titel",
-    a11y_mute: "Stumm schalten oder Ton einschalten", a11y_volume: "Lautstärke", a11y_expand: "Raumsteuerung aufklappen", a11y_collapse: "Raumsteuerung zuklappen",
-    a11y_activate: "{name} bedienen", a11y_set_value: "{name} einstellen", a11y_select_option: "{name} auswählen"
-  },
-  fr: {
-    empty: "Vide", low: "Faible", critical: "Critique", window: "Fenêtre", general: "Général",
-    sensors_manual: "Capteurs (Manuel)", buttons: "Boutons", button: "Bouton", add_button: "Ajouter un bouton",
-    main_climate: "Appareil climatique principal (Optionnel)", climate_info: "Remplit automatiquement Temp/Humidité si vide ci-dessous.",
-    temp_label: "Température (remplace climat)", target_temp_label: "Température cible", humid_label: "Humidité (remplace climat)", temp_unit: "Unité de température", temp_unit_default: "Valeur par défaut de Home Assistant",
-    window_label: "Fenêtres (Liste)", battery_label: "Batteries (Liste)", name: "Nom", icon: "Icône", color: "Couleur",
-    humid_warn_threshold: "Seuil d'alerte d'humidité (%)", high_humidity: "Humidité élevée", device_unavailable: "Appareil indisponible",
-    force_color: "Forcer la couleur", img_url: "URL de l'image", image: "Image", path: "Chemin (Tap Action)", entity: "Entité", device: "Appareil (Optionnel)",
-    image_entity: "Entité lumineuse (Niveaux de gris si éteint)", image_entity_help: "Lorsque l'entité sélectionnée est éteinte, l'image d'en-tête passe en niveaux de gris.",
-    show_image: "Afficher l'image de fond",
-    image_preset_title: "Image de pièce intégrée", image_preset_help: "Choisissez une image RoomCard ou continuez à utiliser votre propre URL ou import.", image_preset_custom: "Image personnalisée",
-    image_preset_living_room: "Salon", image_preset_kitchen: "Cuisine", image_preset_bedroom: "Chambre", image_preset_bathroom: "Salle de bain",
-    image_preset_dining_room: "Salle à manger", image_preset_home_office: "Bureau", image_preset_childrens_room: "Chambre d'enfant", image_preset_hallway: "Entrée",
-    image_preset_guest_room: "Chambre d'amis", image_preset_garage: "Garage", image_preset_garden_patio: "Jardin / terrasse", image_preset_balcony: "Balcon",
-    image_preset_basement: "Sous-sol", image_preset_laundry_room: "Buanderie", image_preset_attic: "Grenier", image_preset_workshop: "Atelier",
-    presence_sensor: "Capteur de présence (Mouvement/Personne)", presence_detected: "Présent", presence_chip_color: "Couleur du badge de présence", presence_solid_background: "Arrière-plan plein du badge de présence",
-    area_setup: "Configuration de zone", area_setup_desc: "Remplir automatiquement les contrôles et capteurs à partir des entités de la zone", area_picker: "Zone Home Assistant", area_generate: "Générer depuis la zone", area_no_entities: "Aucune entité trouvée dans cette zone",
-    alert_sensors: "Capteurs d'alerte", alert_sensor_add: "Ajouter un capteur d'alerte", alert_sensor_entity: "Capteur", alert_sensor_above: "Supérieur à", alert_sensor_below: "Inférieur à", alert_sensor_state: "État", alert_border_color: "Couleur du contour", alert_chip_collapsed: "Afficher en badge groupé", active_alerts: "Alertes actives",
-    template: "Filtre de type", add_template: "avec filtre", add_prefix: "Ajouter",
-    quick_add_title: "Ajout rapide",
-    quick_add_desc: "Ajouter rapidement des boutons à partir d’entités existantes.",
-    quick_add_entity_type_label: "Type d’entité",
-    quick_add_entity_label: "Entité (filtrée)",
-    quick_add_entity_type_help: "Filtre la liste des entités à droite.",
-    quick_add_entity_help: "Sélectionnez une entité de ce type puis cliquez sur « Ajouter ».",
-    quick_add_entity_placeholder: "Sélectionner une entité ({type})",
-    quick_add_empty_hint: "Aucune entité correspondante trouvée.",
-    quick_add_settings_caption: "Paramètres des boutons (pour les éléments ajoutés)",
-    expand_all: "Tout développer",
-    collapse_all: "Tout replier",
-    label_position: "Position du libellé et de l’état",
-    label_position_all: "Position du libellé et de l’état (tous les boutons)",
-    use_global: "Utiliser le réglage global",
-    pos_right: "Droite",
-    pos_bottom: "Bas",
-    pos_left: "Gauche",
-    pos_top: "Haut",
-    row_type: "Type de ligne", type_entity: "Entité", type_template: "Template",
-    tmpl_content: "Contenu (Template)", tmpl_icon: "Icône (Template)", tmpl_color: "Couleur (Template)", tmpl_state: "État (Template)", tmpl_preview: "Aperçu",
-    tmpl_light: "Lumière", tmpl_switch: "Interrupteur / Prise", tmpl_select: "Sélection", tmpl_climate: "Climatisation", tmpl_cover: "Volet / Store", tmpl_media: "Lecteur multimédia",
-    show_state: "Afficher l’état", show_label: "Afficher le libellé", show_icon: "Afficher l’icône", show_last_changed: "Dernier changement", show_sparkline: "Afficher la Sparkline", sparkline_detail: "Ouvrir les détails de l’historique", sparkline_hours: "Historique (heures)", sparkline_refresh: "Actualisation de la Sparkline (secondes)", sparkline_refresh_adjusted: "Saisissez 60 à 3600 secondes. La valeur a été ajustée à {value}.", sparkline_detail_title: "Détails de l’historique", sparkline_loading: "Chargement de l’historique…", sparkline_empty: "Aucune donnée d’historique disponible.", sparkline_error: "Impossible de charger l’historique.", sparkline_current: "Actuel", sparkline_min: "Min", sparkline_max: "Max", sparkline_average: "Moyenne", lc_just_now: "à l’instant", state_first: "Valeur d’abord", text_layout: "Ordre du texte", primary_text: "Première ligne", primary_state: "Valeur d’abord", primary_name: "Nom d’abord",
-    height: "Hauteur", width: "Largeur", align: "Alignement", visible: "Visible", left: "Gauche", center: "Centre", right: "Droite",
-    tap_action: "Appui court", hold_action: "Appui long", double_tap_action: "Double appui",
-    actions: "Actions",
-    act_more: "Détails (Défaut)", act_toggle: "Basculer", act_navigate: "Navigation", act_call_service: "Action (service)", act_none: "Rien",
-    service: "Service (domain.service)", service_data: "Données du service (JSON)", target_entity: "Entité cible",
-    live_preview: "Aperçu en direct",
-    upload_btn: "Télécharger une image", uploading: "Téléchargement...", upload_success: "Image importée", upload_optimized: "Image optimisée et importée",
-    upload_unsupported: "Choisissez une image JPEG, PNG ou WebP.", upload_too_large: "L'image source dépasse la limite de 20 Mo.", upload_decode_error: "L'image n'a pas pu être décodée.", upload_failed: "Échec de l'import ({status}).",
-    image_focal_point: "Point focal de l'image", image_focal_help: "Cliquez ou déplacez le repère vers la zone importante.", image_center: "Centrer", image_horizontal: "Position horizontale", image_vertical: "Position verticale",
-    show_name: "Afficher le titre", header_badges: "Infos d'en-tête supplémentaires", badge_add: "Ajouter une entrée", badge_label: "Libellé (optionnel)", badge_background: "Arrière-plan (rgba)", standard_badge_background: "Fond du badge climat principal (rgba)", badge_auto_climate_btn: "Ajouter automatiquement un bouton de commande du chauffage",
-    visibility: "Visibilité", visibility_cond: "Visibilité conditionnelle", vis_entity: "Entité de la condition", vis_state: "Afficher si l’état est", vis_invert: "Inverser la logique (masquer si l’état correspond)",
-    migration_title: "Action requise",
-    migration_text: "Carte renommée en <b>oneline-room-card</b> pour éviter les conflits.<br>Veuillez changer <code>type: custom:room-card</code> en <code>type: custom:oneline-room-card</code>.",
-    control_mode: "Mode de contrôle", ctrl_default: "Défaut", ctrl_slider: "Curseur", ctrl_buttons: "Boutons", ctrl_full: "Contrôles complets", ctrl_all_options: "Toutes les options",
-    slider_mode: "Mode Curseurs", slider_mode_brightness: "Luminosité", slider_mode_color_temp: "Température de couleur",
-    slider_style: "Style de curseur", style_inline: "Intégré", style_bg: "Arrière-plan",
-    collapsible: "Rétractable", default_state: "État par défaut", state_expanded: "Déplié", state_collapsed: "Replié",
-    header_height: "Hauteur de l'en-tête (px)",
-    typography: "Typographie de l'en-tête", name_font: "Police du nom", info_font: "Police des infos", show_header_text_shadow: "Ombre du texte de l’en-tête",
-    font_size: "Taille (px)", font_weight: "Poids", font_style: "Style", font_color: "Couleur", badge_bg: "Fond du badge",
-    card_behavior: "Comportement de la carte", behavior: "Mode repli", behavior_fixed: "Désactivé", behavior_collapsed: "Replié", behavior_expanded: "Déplié", behavior_remember: "Remember",
-    header: "En-tête", configuration: "Configuration",
-    color_map: "Couleurs par état", color_map_add: "Ajouter couleur", color_map_state: "État",
-    window_always_show: "Toujours afficher (incl. fermé)", window_open_color: "Couleur ouvert", window_closed_color: "Couleur fermé",
-    window_solid_background: "Arrière-plan plein", window_labels: "Libellés des fenêtres",
-    window_custom_label: "Libellé personnalisé",
-    window_open_states: "États ouverts (séparés par virgule)", window_state_colors: "Couleurs par état", window_state_colors_add: "Ajouter couleur",
-    sensors: "Capteurs", show_chip_shadow: "Ombre des badges de capteur", show_status_border: "Contour et lueur d’état", show_status_border_help: "Contrôle uniquement le contour et la lueur d’alerte ; les badges restent visibles.",
-    icon_size: "Taille icône", global_icon_size: "Taille icône globale (px)",
-    header_info_offset: "Position ligne info",
-    header_name_offset: "Position titre",
-    header_sync_offsets: "Synchroniser les positions",
-    global_button_bg: "Fond du bouton global",
-    button_bg: "Fond du bouton",
-    show_cover_presets: "Préréglages de position",
-    cover_presets_label: "Valeurs (séparées par virgule)",
-    show_climate_presets: "Préréglages de température",
-    climate_presets_label: "Températures (séparées par virgule)",
-    show_hvac_modes: "Boutons mode HVAC", show_fan_modes: "Boutons vitesse ventilateur",
-    show_brightness_presets: "Préréglages de luminosité",
-    brightness_presets_label: "Luminosités (séparées par virgule)",
-    show_brightness_value: "Afficher luminosité %",
-    show_color_favorites: "Couleurs favorites",
-    color_favorites_label: "Couleurs ('#hex' ou 'r,g,b', virgule)",
-    show_media_sources: "Sources",
-    show_media_sound_modes: "Modes audio",
-    show_media_title: "Titre média",
-    sub_chips: "Sous-puces", chip_add: "Ajouter une puce", chip_entity: "Entité", chip_attribute: "Attribut (facultatif)", chip_icon: "Icône (facultative)", chip_label: "Libellé (facultatif)", chips_position: "Position des puces", chips_top: "Au-dessus du titre", chips_bottom: "Sous le titre",
-    vis_add: "Ajouter une condition", vis_eq: "L’état est égal à", vis_neq: "L’état est différent de", vis_above: "L’état est strictement supérieur à", vis_below: "L’état est strictement inférieur à",
-    info_line_position: "Position ligne info", info_position_header: "Dans l'en-tête (défaut)", info_position_below: "Sous l'en-tête",
-    last_activity_title: "Dernière activité", last_activity_show: "Afficher la dernière activité",
-    room_modes: "Modes de pièce", room_modes_help: "Lancer des scènes ou des scripts depuis la carte de pièce.", room_mode_add: "Ajouter un mode", room_mode_entity: "Scène ou script", room_mode_active_when: "Actif lorsque", room_mode_remove: "Supprimer le mode", room_mode_up: "Déplacer vers le haut", room_mode_down: "Déplacer vers le bas",
-    adaptive_images: "Images adaptatives", adaptive_images_help: "La première règle correspondante remplace l’image d’en-tête par défaut.", adaptive_image_add: "Ajouter une règle d’image", adaptive_image_name: "Nom de la règle (facultatif)", adaptive_image_conditions: "Conditions", adaptive_image_preset: "Image intégrée", adaptive_image_custom: "URL personnalisée / import", adaptive_image_duplicate: "Dupliquer la règle d’image", adaptive_image_remove: "Supprimer la règle d’image", adaptive_image_up: "Monter la règle d’image", adaptive_image_down: "Descendre la règle d’image", adaptive_image_position: "Point focal de la règle",
-    status_groups: "État de la pièce", status_groups_help: "Résumer les entités configurées sans modifier la priorité des alertes.", status_group_add: "Ajouter un groupe d’état", status_group_preset_lights: "Lumières allumées", status_group_preset_windows: "Fenêtres ouvertes", status_group_preset_media: "Média actif", status_group_preset_power: "Puissance", status_group_entities: "Entités", status_group_active_states: "États actifs (séparés par des virgules)", status_group_display: "Affichage", status_group_count: "Nombre", status_group_value: "Somme numérique", status_group_unit: "Unité de sortie", status_group_precision: "Décimales", status_group_hide_zero: "Masquer à zéro", status_group_details: "Ouvrir les entités contributrices", status_group_conditions: "Afficher le groupe lorsque", status_group_duplicate: "Dupliquer le groupe d’état", status_group_remove: "Supprimer le groupe d’état", status_group_up: "Monter le groupe d’état", status_group_down: "Descendre le groupe d’état", status_group_incompatible_units: "Unités incompatibles",
-    a11y_close: "Fermer", a11y_previous_track: "Piste précédente", a11y_play_pause: "Lecture ou pause", a11y_next_track: "Piste suivante",
-    a11y_mute: "Couper ou rétablir le son", a11y_volume: "Volume", a11y_expand: "Développer les commandes", a11y_collapse: "Réduire les commandes",
-    a11y_activate: "Activer {name}", a11y_set_value: "Régler {name}", a11y_select_option: "Sélectionner {name}"
-  }
-};
-
-const getTranslation = (hass, key) => {
-  const lang = hass?.language?.split("-")[0] || "en";
-  return TRANSLATIONS[lang]?.[key] ?? TRANSLATIONS["en"]?.[key] ?? key;
-};
